requirements.txt
streamlit
google.generativeai
python_dotenv